Move NoScriptViewer to its own component
